update and move util functions to util.py
