+++ generative/maisi/configs/config_maisi_controlnet_train.json
@@ -3,7 +3,7 @@
     "spatial_dims": 3,
     "image_channels": 1,
     "latent_channels": 4,
-    "difusion_unet_def": {
+    "diffusion_unet_def": {
         "_target_": "monai.apps.generation.maisi.networks.diffusion_model_unet_maisi.DiffusionModelUNetMaisi",
         "spatial_dims": "@spatial_dims",
         "in_channels": "@latent_channels",
