+++ generative/maisi/scripts/train_controlnet.py
@@ -10,172 +10,23 @@
 # limitations under the License.
 
 import argparse
-import copy
 import json
 import logging
 import os
 import sys
 import time
-from argparse import Namespace
 from datetime import timedelta
 from pathlib import Path
-from typing import Any
 
 import torch
 import torch.distributed as dist
 import torch.nn.functional as F
-from monai.bundle import ConfigParser
-from monai.data import CacheDataset, DataLoader, partition_dataset
 from monai.networks.utils import copy_model_state
-from monai.transforms import Compose, EnsureTyped, Lambdad, LoadImaged, Orientationd
 from monai.utils import RankFilter
 from torch.cuda.amp import GradScaler, autocast
 from torch.nn.parallel import DistributedDataParallel as DDP
 from torch.utils.tensorboard import SummaryWriter
-from utils import binarize_labels
-
-
-def setup_ddp(rank: int, world_size: int) -> torch.device:
-    """Initialize the distributed process group.
-
-    Args:
-        rank (int): rank of the current process.
-        world_size (int): number of processes participating in the job.
-    """
-    dist.init_process_group(
-        backend="nccl", init_method="env://", timeout=timedelta(seconds=36000), rank=rank, world_size=world_size
-    )
-    dist.barrier()
-    device = torch.device(f"cuda:{rank}")
-    return device
-
-
-def define_instance(args: Namespace, instance_def_key: str) -> Any:
-    """Get the parsed instance based on provided attributes.
-
-    Args:
-        args (Namespace): a object for storing attributes.
-        instance_def_key (str): key associated to target instance.
-
-    Returns:
-        Any: parsed instance.
-    """
-    parser = ConfigParser(vars(args))
-    parser.parse(True)
-    return parser.get_parsed_content(instance_def_key, instantiate=True)
-
-
-def add_data_dir2path(list_files: list, data_dir: str, fold: int = None) -> tuple[list, list]:
-    """Read a list of data dictionary.
-
-    Args:
-        list_files (list): input data to load and transform to generate dataset for model.
-        data_dir (str): directory of files.
-        fold (int, optional): fold index for cross validation. Defaults to None.
-
-    Returns:
-        tuple[list, list]: A tuple of two arrays (training, validation).
-    """
-    new_list_files = copy.deepcopy(list_files)
-    if fold is not None:
-        new_list_files_train = []
-        new_list_files_val = []
-    for d in new_list_files:
-        d["image"] = os.path.join(data_dir, d["image"])
-
-        if "label" in d:
-            d["label"] = os.path.join(data_dir, d["label"])
-
-        if fold is not None:
-            if d["fold"] == fold:
-                new_list_files_val.append(copy.deepcopy(d))
-            else:
-                new_list_files_train.append(copy.deepcopy(d))
-
-    if fold is not None:
-        return new_list_files_train, new_list_files_val
-    else:
-        return new_list_files, []
-
-
-def prepare_maisi_controlnet_json_dataloader(
-    json_data_list: list | str,
-    data_base_dir: list | str,
-    batch_size: int = 1,
-    fold: int = 0,
-    cache_rate: float = 0.0,
-    rank: int = 0,
-    world_size: int = 1,
-) -> tuple[DataLoader, DataLoader]:
-    """Prepare dataloaders for training and validation.
-
-    Args:
-        json_data_list (list | str): the name of JSON files listing the data.
-        data_base_dir (list | str): directory of files.
-        batch_size (int, optional): how many samples per batch to load . Defaults to 1.
-        fold (int, optional): fold index for cross validation. Defaults to 0.
-        cache_rate (float, optional): percentage of cached data in total. Defaults to 0.0.
-        rank (int, optional): rank of the current process. Defaults to 0.
-        world_size (int, optional): number of processes participating in the job. Defaults to 1.
-
-    Returns:
-        tuple[DataLoader, DataLoader]:  A tuple of two dataloaders (training, validation).
-    """
-    ddp_bool = world_size > 1
-    if isinstance(json_data_list, list):
-        assert isinstance(data_base_dir, list)
-        list_train = []
-        list_valid = []
-        for data_list, data_root in zip(json_data_list, data_base_dir):
-            with open(data_list, "r") as f:
-                json_data = json.load(f)["training"]
-            train, val = add_data_dir2path(json_data, data_root, fold)
-            list_train += train
-            list_valid += val
-    else:
-        with open(json_data_list, "r") as f:
-            json_data = json.load(f)["training"]
-        list_train, list_valid = add_data_dir2path(json_data, data_base_dir, fold)
-
-    common_transform = [
-        LoadImaged(keys=["image", "label"], image_only=True, ensure_channel_first=True),
-        Orientationd(keys=["label"], axcodes="RAS"),
-        EnsureTyped(keys=["label"], dtype=torch.uint8, track_meta=True),
-        Lambdad(keys="top_region_index", func=lambda x: torch.FloatTensor(x)),
-        Lambdad(keys="bottom_region_index", func=lambda x: torch.FloatTensor(x)),
-        Lambdad(keys="spacing", func=lambda x: torch.FloatTensor(x)),
-        Lambdad(keys="top_region_index", func=lambda x: x * 1e2),
-        Lambdad(keys="bottom_region_index", func=lambda x: x * 1e2),
-        Lambdad(keys="spacing", func=lambda x: x * 1e2),
-    ]
-    train_transforms, val_transforms = Compose(common_transform), Compose(common_transform)
-
-    train_loader = None
-
-    if ddp_bool:
-        list_train = partition_dataset(
-            data=list_train,
-            shuffle=True,
-            num_partitions=world_size,
-            even_divisible=True,
-        )[rank]
-    train_ds = CacheDataset(data=list_train, transform=train_transforms, cache_rate=cache_rate, num_workers=8)
-    train_loader = DataLoader(train_ds, batch_size=batch_size, shuffle=True, num_workers=8, pin_memory=True)
-    if ddp_bool:
-        list_valid = partition_dataset(
-            data=list_valid,
-            shuffle=True,
-            num_partitions=world_size,
-            even_divisible=False,
-        )[rank]
-    val_ds = CacheDataset(
-        data=list_valid,
-        transform=val_transforms,
-        cache_rate=cache_rate,
-        num_workers=8,
-    )
-    val_loader = DataLoader(val_ds, batch_size=batch_size, shuffle=False, num_workers=2, pin_memory=False)
-    return train_loader, val_loader
+from utils import binarize_labels, define_instance, prepare_maisi_controlnet_json_dataloader, setup_ddp
 
 
 def main():
@@ -198,8 +49,8 @@ def main():
     # Step 0: configuration
     logger = logging.getLogger("maisi.controlnet.training")
     # whether to use distributed data parallel
-    ddp_bool = args.gpus > 1
-    if ddp_bool:
+    use_ddp = args.gpus > 1
+    if use_ddp:
         rank = int(os.environ["LOCAL_RANK"])
         world_size = int(os.environ["WORLD_SIZE"])
         device = setup_ddp(rank, world_size)
@@ -240,8 +91,10 @@ def main():
 
     # Step 2: define diffusion model and controlnet
     # define diffusion Model
-    unet = define_instance(args, "difusion_unet_def").to(device)
+    unet = define_instance(args, "diffusion_unet_def").to(device)
     # load trained diffusion model
+    if not os.path.exists(args.trained_diffusion_path):
+        raise ValueError("Please download the trained diffusion unet checkpoint.")
     diffusion_model_ckpt = torch.load(args.trained_diffusion_path, map_location=device)
     unet.load_state_dict(diffusion_model_ckpt["unet_state_dict"])
     # load scale factor
@@ -266,7 +119,7 @@ def main():
 
     noise_scheduler = define_instance(args, "noise_scheduler")
 
-    if ddp_bool:
+    if use_ddp:
         controlnet = DDP(controlnet, device_ids=[device], output_device=rank, find_unused_parameters=True)
 
     # Step 3: training config
@@ -377,7 +230,7 @@ def main():
 
         epoch_loss = epoch_loss_ / (step + 1)
 
-        if ddp_bool:
+        if use_ddp:
             dist.barrier()
             dist.all_reduce(epoch_loss, op=torch.distributed.ReduceOp.AVG)
 
@@ -407,7 +260,7 @@ def main():
                 )
 
         torch.cuda.empty_cache()
-    if ddp_bool:
+    if use_ddp:
         dist.destroy_process_group()
